+++ tdd/src/main/java/com/work/tdd/euler/util/Matrix.java
@@ -5,24 +5,24 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 public class Matrix {
 
-    public static Long[] matrixSolution(Long[][] inputs, Long[] outputs) {
+    public static Double[] matrixSolution(Double[][] inputs, Double[] outputs) {
         checkArgument(inputs.length == outputs.length);
-        Long[] solution = new Long[outputs.length];
-        long originalDeterminant = determinant(inputs);
+        Double[] solution = new Double[outputs.length];
+        Double originalDeterminant = determinant(inputs);
         for (int i = 0; i < outputs.length; i++) {
-            Long[][] replacedMatrix = replaceColumn(inputs, outputs, i);
+            Double[][] replacedMatrix = replaceColumn(inputs, outputs, i);
             solution[i] = determinant(replacedMatrix) / originalDeterminant;
         }
         return solution;
     }
 
-    public static long determinant(Long[][] matrix) {
+    public static Double determinant(Double[][] matrix) {
         checkNotNull(matrix);
         checkArgument(matrix.length == matrix[0].length);
         if (matrix.length == 1) {
             return matrix[0][0];
         }
-        long determinant = 0;
+        Double determinant = 0d;
         int sign = 1;
         for (int row = 0; row < matrix.length; row++) {
             determinant += matrix[row][0] * sign * determinant(subMatrix(matrix, row));
@@ -31,8 +31,8 @@ public class Matrix {
         return determinant;
     }
 
-    private static Long[][] subMatrix(Long[][] matrix, int chosenRow) {
-        Long[][] subMatrix = new Long[matrix.length - 1][matrix.length - 1];
+    private static Double[][] subMatrix(Double[][] matrix, int chosenRow) {
+        Double[][] subMatrix = new Double[matrix.length - 1][matrix.length - 1];
         for (int row = 0; row < matrix.length; row++) {
             if (row != chosenRow) {
                 for (int column = 1; column < matrix.length; column++) {
@@ -44,8 +44,8 @@ public class Matrix {
         return subMatrix;
     }
 
-    private static Long[][] replaceColumn(Long[][] matrix, Long[] newColumn, Integer targetColumn) {
-        Long[][] replacedMatrix = new Long[matrix.length][matrix.length];
+    private static Double[][] replaceColumn(Double[][] matrix, Double[] newColumn, Integer targetColumn) {
+        Double[][] replacedMatrix = new Double[matrix.length][matrix.length];
         for (int row = 0; row < matrix.length; row++) {
             for (int column = 0; column < matrix.length; column++) {
                 replacedMatrix[row][column] = (column == targetColumn) ? newColumn[row] : matrix[row][column];
